pyproject.toml
[project]
name = "http-sse-mcp-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "html2text>=2024.2.26",
    "mcp[cli]>=1.2.0",
    "requests>=2.32.3",
    "streamlit>=1.44.1",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "pydantic>=2.5.0",
]
